requirements.txt
flask==3.0.0
flask-cors==4.0.0
transformers==4.36.0
torch==2.1.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
tensorflow==2.15.0
tokenizers==0.15.0